Expand the introduction to talk about a wider range of problems
We discussed on the list and elsewhere a borader set of cases where
something like WRAP_UP could be useful. The specifics are still in flux.
The were some thoughts on other capsule types we could define, or we may
decide to punt on that work entirely.

However, I thought it might be useful to put into prose (my recollection)
of the discussion of the problem space.

I left it as a "read between the lines" thing, but I have use cases where
wrapping HTTP streams in a capsule protocol can help address operational
issues within a system. I'm hoping moving away from CONNECT proxy specific
language can help folks realise there's other applicability.

This is wall of text. We might decide that cherry-picking pieces out
is the best thing for now.